+++ src/rules/font-weight-notation/__tests__/index.js
@@ -21,6 +21,9 @@ testRule("numeric", tr => {
   tr.ok("a { font: italic small-caps 400 16px/500 cursive; }")
   tr.ok("a { font: italic small-caps 400 16px/500 \"bold font name\"; }")
   tr.ok("a { font: italic small-caps 400 16px/500 boldfontname; }")
+  tr.ok("a { font: normal 400 normal 16px serif; }", "two normals and a numbered weight")
+  tr.ok("a { font: 400 normal 16px serif; }", "one normal and a numbered weight")
+  tr.ok("a { font: 400 16px serif; }", "no normals and a numbered weight")
 
   tr.notOk("a { font-weight: normal; }", {
     message: messages.expected("numeric"),
@@ -32,6 +35,21 @@ testRule("numeric", tr => {
     line: 1,
     column: 29,
   })
+  tr.notOk("a { font: normal 16px/3 cursive; }", {
+    message: messages.expected("numeric"),
+    line: 1,
+    column: 11,
+  }, "one normal and no numbered weight")
+  tr.notOk("a { font: normal normal 16px/3 cursive; }", {
+    message: messages.expected("numeric"),
+    line: 1,
+    column: 11,
+  }, "two normals and no numbered weight")
+  tr.notOk("a { font: normal normal normal 16px/3 cursive; }", {
+    message: messages.expected("numeric"),
+    line: 1,
+    column: 11,
+  }, "three normals and no numbered weight")
 })
 
 testRule("numeric", { ignore: ["relative"] }, tr => {
@@ -63,6 +81,9 @@ testRule("named", tr => {
   tr.ok("a { font: italic small-caps bold 16px/500 cursive; }")
   tr.ok("a { font: italic small-caps bold 16px/500 \"cursive 100 font\"; }")
   tr.ok("a { font: italic small-caps bold 16px/500 100cursivefont; }")
+  tr.ok("a { font: normal 16px/3 cursive; }", "one normal and no numbered weight")
+  tr.ok("a { font: normal normal 16px/3 cursive; }", "two normals and no numbered weight")
+  tr.ok("a { font: normal normal normal 16px/3 cursive; }", "three normals and no numbered weight")
 
   tr.notOk("a { font-weight: 400; }", {
     message: messages.expected("named"),
@@ -79,4 +100,19 @@ testRule("named", tr => {
     line: 1,
     column: 29,
   })
+  tr.notOk("a { font: normal 400 normal 16px serif; }", {
+    message: messages.expected("named"),
+    line: 1,
+    column: 18,
+  }, "two normals and a numbered weight")
+  tr.notOk("a { font: 400 normal 16px serif; }", {
+    message: messages.expected("named"),
+    line: 1,
+    column: 11,
+  }, "one normal and a numbered weight")
+  tr.notOk("a { font: 400 16px serif; }", {
+    message: messages.expected("named"),
+    line: 1,
+    column: 11,
+  }, "no normals and a numbered weight")
 })