+++ src/rules/font-weight-notation/index.js
@@ -1,4 +1,5 @@
 import postcss from "postcss"
+import { includes } from "lodash"
 import {
   cssWordIsVariable,
   declarationValueIndexOffset,
@@ -15,7 +16,8 @@ export const messages = ruleMessages(ruleName, {
   invalidNamed: name => `Unexpected invalid font-weight name "${name}"`,
 })
 
-const NAMED_WEIGHTS = [ "normal", "bold", "bolder", "lighter" ]
+const WEIGHT_SPECIFIC_KEYWORDS = [ "bold", "bolder", "lighter" ]
+const NORMAL_KEYWORD = "normal"
 const RELATIVE_NAMED_WEIGHTS = [ "bolder", "lighter" ]
 function isNumbery(x) {
   return Number(x) == x
@@ -46,11 +48,18 @@ export default function (expectation, options) {
     })
 
     function checkFont(decl) {
+      const valueList = postcss.list.space(decl.value)
+      // We do not need to more carefully distinguish font-weight
+      // numbers from unitless line-heights because line-heights in
+      // `font` values need to be part of a font-size/line-height pair
+      const hasNumericFontWeight = valueList.some(isNumbery)
+
       for (let value of postcss.list.space(decl.value)) {
-        // We do not need to more carefully distinguish font-weight
-        // numbers from unitless line-heights because line-heights in
-        // `font` values need to be part of a font-size/line-height pair
-        if (isNumbery(value) || NAMED_WEIGHTS.indexOf(value) !== -1) {
+        if (
+          (value === NORMAL_KEYWORD && !hasNumericFontWeight)
+          || isNumbery(value)
+          || includes(WEIGHT_SPECIFIC_KEYWORDS, value)
+        ) {
           checkWeight(value, decl)
           return
         }
@@ -60,7 +69,7 @@ export default function (expectation, options) {
     function checkWeight(weightValue, decl) {
       if (cssWordIsVariable(weightValue)) { return }
       if (optionsHaveIgnored(options, "relative") &&
-        RELATIVE_NAMED_WEIGHTS.indexOf(weightValue) !== -1) { return }
+        includes(RELATIVE_NAMED_WEIGHTS, weightValue)) { return }
 
       const weightValueOffset = decl.value.indexOf(weightValue)
 
@@ -74,7 +83,7 @@ export default function (expectation, options) {
         if (isNumbery(weightValue)) {
           return complain(messages.expected("named"))
         }
-        if (NAMED_WEIGHTS.indexOf(weightValue) === -1) {
+        if (!includes(WEIGHT_SPECIFIC_KEYWORDS, weightValue) && weightValue !== NORMAL_KEYWORD) {
           return complain(messages.invalidNamed(weightValue))
         }
         return